move didv plotting to core/didv, turn them into a class
