+++ qetpy/plotting/__init__.py
@@ -1,7 +1,5 @@
-from ._didv_plotting import *
 from ._fitting_plotting import *
 from ._iv_plotting import *
 from ._noise_plotting import *
 from ._of_nsmb_plotting import *
 from ._ibis_plotting import *
-from ._didv2_plotting import *
